refactor: extract models.go
